moved ddcio to the local cmd tree
