+++ cmd/local/apicollect/jobprofiles.go
@@ -22,11 +22,11 @@ import (
 	"path/filepath"
 
 	"github.com/dremio/dremio-diagnostic-collector/cmd/local/conf"
+	"github.com/dremio/dremio-diagnostic-collector/cmd/local/ddcio"
 	"github.com/dremio/dremio-diagnostic-collector/cmd/local/queriesjson"
 	"github.com/dremio/dremio-diagnostic-collector/cmd/local/restclient"
 	"github.com/dremio/dremio-diagnostic-collector/cmd/local/threading"
 	"github.com/dremio/dremio-diagnostic-collector/cmd/simplelog"
-	"github.com/dremio/dremio-diagnostic-collector/pkg/ddcio"
 )
 
 func getNumberOfJobProfilesCollected(c *conf.CollectConf) (tried, collected int, err error) {
