+++ cmd/local/apicollect/kvreports.go
@@ -22,9 +22,9 @@ import (
 	"path/filepath"
 
 	"github.com/dremio/dremio-diagnostic-collector/cmd/local/conf"
+	"github.com/dremio/dremio-diagnostic-collector/cmd/local/ddcio"
 	"github.com/dremio/dremio-diagnostic-collector/cmd/local/restclient"
 	"github.com/dremio/dremio-diagnostic-collector/cmd/simplelog"
-	"github.com/dremio/dremio-diagnostic-collector/pkg/ddcio"
 )
 
 func RunCollectKvReport(c *conf.CollectConf) error {
