+++ cmd/local/conf/autodetect/gclog_finder.go
@@ -22,7 +22,7 @@ import (
 	"path"
 	"strings"
 
-	"github.com/dremio/dremio-diagnostic-collector/pkg/ddcio"
+	"github.com/dremio/dremio-diagnostic-collector/cmd/local/ddcio"
 )
 
 // findGCLogLocation retrieves the gc log location with a search string to greedily retrieve everything by prefix
