+++ cmd/local/ddcio/testdata/file1_copy.txt
