+++ cmd/local/logcollect/logcollect.go
@@ -24,8 +24,8 @@ import (
 	"strings"
 	"time"
 
+	"github.com/dremio/dremio-diagnostic-collector/cmd/local/ddcio"
 	"github.com/dremio/dremio-diagnostic-collector/cmd/simplelog"
-	"github.com/dremio/dremio-diagnostic-collector/pkg/ddcio"
 )
 
 type Collector struct {
